+++ RustDMA/SDK/Item.h
@@ -3,7 +3,7 @@
 class Item
 {
 	uint64_t Class = 0x0;
-	uint64_t ItemID = 0x20; // public ItemId uid;
+	uint32_t ItemID = 0x20; // public ItemId uid;
 	uint64_t HeldEntity = 0xB0; //private EntityRef heldEntity;
 	BaseProjectile* BaseProjectileInstance = nullptr;
 public:
